# docs/swagger.yaml
basePath: /
definitions:
  model.Person:
    properties:
      ID:
        type: string
      married:
        type: boolean
      profession:
        maxLength: 30
        minLength: 3
        type: string
      salary:
        maximum: 100000
        minimum: 100
        type: integer
    required:
    - profession
    - salary
    type: object
  model.PersonRequest:
    properties:
      married:
        type: boolean
      profession:
        maxLength: 30
        minLength: 3
        type: string
      salary:
        maximum: 100000
        minimum: 100
        type: integer
    required:
    - profession
    - salary
    type: object
  model.RefreshRequest:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  model.UserRequest:
    properties:
      password:
        maxLength: 15
        minLength: 4
        type: string
      username:
        maxLength: 15
        minLength: 4
        type: string
    required:
    - password
    - username
    type: object
  service.TokenPair:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing persons and users
  title: FirstTask API
  version: "1.0"
paths:
  /downloadImage/{imageName}:
    get:
      consumes:
      - application/json
      description: Downloads the specified image from the server
      parameters:
      - description: Image filename
        in: path
        name: imageName
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Image file
          schema:
            type: file
        "404":
          description: Not Found
          schema: {}
      summary: Download an image
      tags:
      - Images
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with the provided username and password
      parameters:
      - description: userRequest value (model.PersonRequest)
        in: body
        name: userRequest
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "400":
          description: Bad Request
          schema: {}
      summary: User login
      tags:
      - User
  /persons:
    get:
      consumes:
      - application/json
      description: Get all persons
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Person'
            type: array
        "400":
          description: Bad Request
          schema: {}
      summary: Get all persons
      tags:
      - Person
    post:
      consumes:
      - application/json
      description: Creates a new person
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      - description: personRequest value (model.PersonRequest)
        in: body
        name: personRequest
        required: true
        schema:
          $ref: '#/definitions/model.PersonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Person'
        "400":
          description: Bad Request
          schema: {}
      summary: Create a new person
      tags:
      - Person
  /persons/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a person by ID
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      - description: Person ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a person by ID
      tags:
      - Person
    get:
      consumes:
      - application/json
      description: Get a person by ID
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      - description: Person ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Person'
        "400":
          description: Bad Request
          schema: {}
      summary: Get a person by ID
      tags:
      - Person
    put:
      consumes:
      - application/json
      description: Update a person by ID
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      - description: Person ID
        in: path
        name: id
        required: true
        type: string
      - description: personRequest value (model.PersonRequest)
        in: body
        name: personRequest
        required: true
        schema:
          $ref: '#/definitions/model.PersonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Person'
        "400":
          description: Bad Request
          schema: {}
      summary: Update a person by ID
      tags:
      - Person
  /refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access and refresh tokens using the provided refresh
        token
      parameters:
      - description: refreshRequest value (model.RefreshRequest)
        in: body
        name: refreshRequest
        required: true
        schema:
          $ref: '#/definitions/model.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "400":
          description: Bad Request
          schema: {}
      summary: Refreshes access and refresh tokens
      tags:
      - Authentication
  /signUp:
    post:
      consumes:
      - application/json
      description: Sign up a new user
      parameters:
      - description: userRequest value (model.PersonRequest)
        in: body
        name: userRequest
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
      summary: Sign up a new user
      tags:
      - User
  /uploadImage:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image to the server
      parameters:
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
      summary: Upload an image
      tags:
      - Images
swagger: "2.0"